Move password verification and hash parsing functions into separate file
